Log & Frame views & features
Fixed all these

+++ cleanup.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-mkdir -p var/{lib,log}
-rm -f $MAC_PROV_ROOT/var/lib/*
-rm -f $MAC_PROV_ROOT/var/log/*
-touch $MAC_PROV_ROOT/var/lib/keystore.db
-node $MAC_PROV_ROOT/bin/keystore --init-db
+if ! test -z "$MAC_PROV_ROOT"; then
+  mkdir -p var/{lib,log}
+  rm -f $MAC_PROV_ROOT/var/lib/*
+  rm -f $MAC_PROV_ROOT/var/log/*
+  touch $MAC_PROV_ROOT/var/lib/keystore.db
+  node $MAC_PROV_ROOT/bin/keystore --init-db
+else
+  echo MAC_PROV_ROOT not defined
+fi
